fix(evidence): stop unsafe typed claim promotion
